requirements.txt
anyhttp==0.1
anyio==3.6.2
attrs==22.1.0
bcrypt==4.0.1
black==22.12.0
certifi==2022.12.7
click==8.1.3
fastapi==0.88.0
flake8==6.0.0
h11==0.14.0
httpcore==0.16.2
httptools==0.5.0
httpx==0.23.1
idna==3.4
iniconfig==1.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
nanoid==2.0.0
packaging==22.0
pathspec==0.10.3
platformdirs==2.6.0
pluggy==1.0.0
pycodestyle==2.10.0
pydantic==1.10.2
pyflakes==3.0.1
PyJWT==2.6.0
PyMySQL==1.0.2
pytest==7.2.0
python-dotenv==0.21.0
PyYAML==6.0
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==1.4.45
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
